refactored the code to function with external API
